Move HW factories back
