+++ tests/xharness/xharness.csproj
@@ -76,6 +76,7 @@
     <Compile Include="AppRunner.cs" />
     <Compile Include="BCLTestImportTargetFactory.cs" />
     <Compile Include="CrashSnapshotReporterFactory.cs" />
+    <Compile Include="DeviceLoaderFactory.cs" />
     <Compile Include="DeviceLogCapturerFactory.cs" />
     <Compile Include="GitHub.cs" />
     <Compile Include="Harness.cs" />
@@ -100,6 +101,7 @@
     <Compile Include="MonoNativeInfo.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SimulatorLoaderFactory.cs" />
     <Compile Include="SolutionGenerator.cs" />
     <Compile Include="Targets\iOSTarget.cs" />
     <Compile Include="Targets\MacTarget.cs" />
